Video 31 Deploy FastAPI/requirements.txt
fastapi==0.96.0
uvicorn==0.22.0
python-multipart==0.0.5

numpy==1.22.4
jinja2==3.1.2
tensorflow==2.12.0
pillow==9.0.1
keras==2.12.0